Consolidates Array validation type checking
